requirements.txt
# Core FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
starlette==0.27.0

# Authentication and Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==42.0.0

# Database and Data Handling
pydantic[email]==2.5.0
pydantic-settings==2.1.0
aiosqlite==0.19.0
python-csv==0.0.13

# AI / Machine Learning / LLMs
langchain==0.2.1
langchain-community==0.0.36
langchain-core==0.3.60
langchain-ollama==0.3.3
ollama>=0.4.8,<1.0.0



# PDF Processing
pdfplumber==0.10.0
pypdf2==3.0.1

# Audio Processing
gTTS==2.4.0
git+https://github.com/openai/whisper.git
mutagen==1.47.0

# Language Processing
langid==1.1.6
deep-translator==1.11.4
spacy==3.7.2
nltk==3.8.1

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
psutil==5.9.6
filetype==1.2.0
python-magic==0.4.27
python-dateutil==2.8.2
numpy==1.24.3
pandas==2.1.4
pillow==10.1.0

# Logging and Monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Caching
redis==5.0.1
aioredis==2.0.1

# Performance Optimization
orjson==3.9.10
ujson==5.8.0

# File Compression
zipfile38==0.0.3

# Documentation (Optional)
mkdocs==1.5.3
mkdocs-material==9.4.8

# Development and Testing Tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Packaging Tools
setuptools>=65.0.0
wheel>=0.37.0
pip>=21.0.0
